Заканчивает основную часть дубль 2
