+++ src/basis-constants.js
@@ -1,7 +1,11 @@
 export const basisConstants = {
   arrivals: [`Check-in`, `Sightseeing`, `Restaurant`],
   vehicles: [`Taxi`, `Bus`, `Train`, `Ship`, `Transport`, `Drive`, `Flight`],
-  descriptionText: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.`
+};
+
+export const EvtKey = {
+  ESCAPE: `Escape`,
+  ESC: `Esc`,
 };
 
 export const SortType = {
@@ -65,3 +69,16 @@ export const typeIcons = {
   'sightseeing': `🏛`,
   'restaurant': `🍴`
 };
+
+export const Mode = {
+  DEFAULT: `DEFAULT`,
+  EDITING: `EDITING`
+};
+
+export const State = {
+  SAVING: `SAVING`,
+  DELETING: `DELETING`,
+  ABORTING: `ABORTING`
+};
+
+
